requirements.txt
# Core dependencies
fastapi==0.104.1
uvicorn[standard]==0.24.0
python-dotenv==1.0.0
requests==2.31.0

# Data processing
pandas==2.1.3
numpy==1.26.2

# Additional dependencies
matplotlib==3.8.1
google-generativeai==0.3.0
pydantic==2.5.0
